[Parser] StreamingAsset 파싱할 데이터 추가 및 디렉터리 내 파일들 한번에 파싱

+++ Assets/Scripts/PlaySys/Sound.cs
@@ -26,11 +26,11 @@ public class Sound
 
     public struct Timing
     {
-        public int time;          // time when bpm changes
+        public float time;          // time when bpm changes
         public float bpm;         // 1 / beat length * 1000 ( ms ) * 60 ( minute )
         private float beatLength;
 
-        public Timing( int _time, float _beatLength )
+        public Timing( float _time, float _beatLength )
         {
             time = _time;
             beatLength = _beatLength;
